Added 4 files including tiny Llama model code
The 4 files include 3 python scripts - to check authenticity of text and json files and to append them into a csv (xlsx) file
Another one is a ipynb file that contains the code used to train the tiny llama model on the data. Unfortunately we weren't able to complete training due to lack of GPU resources.